Make compatible with redis.py ~2.7-2.10.
Strangely the rename function threw errors when the first key did
not exists. The most minimal change I could think of was replacing
them with custom Lua rename scripts that just don't care for existing
keys. There is probably a better non-script solution but that would
require (major) changes to the core logic of autoclean.